fix bugs

+++ aaas-api-user-info/src/main/java/com/shellshellfish/aaas/userinfo/dao/repositories/mysql/UserInfoAssetsRepository.java
@@ -5,8 +5,8 @@ import java.math.BigInteger;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface UserInfoAssetsRepository extends PagingAndSortingRepository<UiAsset, BigInteger> {
+public interface UserInfoAssetsRepository extends PagingAndSortingRepository<UiAsset, Long> {
 
-  UiAsset findByUserId(BigInteger userId);
+  UiAsset findByUserId(Long userId);
 
 }
