+++ aaas-api-user-info/src/main/java/com/shellshellfish/aaas/userinfo/dao/repositories/mysql/UserPortfolioRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 public interface UserPortfolioRepository extends PagingAndSortingRepository<UiPortfolio, Long> {
 
-  List<UiPortfolio> findAllByUserId( BigInteger userId);
+  List<UiPortfolio> findAllByUserId( Long userId);
 
 }
